HTML report: use semantic class names instead of bg-<color>
This commit prepares the rework of the colors used in the HTML
reports. It should not change the behavior of the generated HTML page
in any way, it just use semantic class names ("lint-errors") instead
of color-related names ("bg-red") to clarify the purpose, and make it
easier to change the color rendering in the future.