HTML report: remove the semantics distinction between 'green' and 'greeny'
In report summaries (report-last.html as opposed to report-full.html),
'greeny' (light green) was used when all available versions of
a package builds, but some OCaml compiler versions are not available
for this package -- as opposed to the full green used when all
monitored compiler versions are available for this package and they
all build.

This criterion is fragile to addition of new compiler versions, and
suggesting that packages that don't support old OCaml compiler
versions are "not as good" is a rather debatable choice.

This commit removes the distinction between those two states. Either
there is no monitored version that supports the package ("fail"), or
at least one version is available and all available versions build
correctly (full), or only some versions work (partial).