HTML reports: make Fail a "red" error, and de-emphasize Deps (white background)
The real problem that opam-builder is good at solving is builds that
fail: "failure" of package builds. This should be the central issue
that gets the attention-catching color.

On the other hand, there are good reasons why a package could be
uninstallable due to its dependencies, and we should not require
marking a package unavailable on an OCaml version just because some of
its dependencies also are uninstallable -- this is a non-modular
requirement.