allow opening package logs in a new tab
The existing Javascript frontend does not allow to open package logs
in new browser tabs, because it uses a fake URL to link to the logs
that just runs Javascript to update the page state.

This patch uses a real URL instead, leading to a page in the desired
state.